アニメーション速度の変更を実装。ただしバグが残る。
キャラチップを表示していない状態でスピードを変更するとエラー。

+++ src/app/config/ConfigStageController.java
@@ -33,7 +33,7 @@ public class ConfigStageController {
     mainController.updateZoomRate(n.doubleValue());
   });
   animationSpeedSlider.valueProperty().addListener((obs, o, n) -> {
-    mainController.updateAnimationSpeed(n.doubleValue());
+    mainController.updateAnimationSpeed(n.intValue());
   });
 
 }//}}}
